Moved camera files to camera folder
